Fix error on assigning groups to users
Since PHP 8.0 the count() function throws a TypeError when a non-countable type is passed to it. This resulted in the following error when trying to change the group assignment of a user:
```
PHP Fatal error:  Uncaught TypeError: count():
  Argument #1 ($value) must be of type Countable|array, string given
  in usvn/library/Zend/Db/Table/Abstract.php:1236

Stack trace:
#0 usvn/library/USVN/Db/Table/Row/User.php(31): Zend_Db_Table_Abstract->find()
#1 usvn/library/USVN/User.php(118): USVN_Db_Table_Row_User->addGroup()
#2 usvn/app/controllers/UseradminController.php(125): USVN_User->save()
...
``` 

The Zend code already coerces the passed argument to an array, unfortunately only after the call to count(). To fix this we just move it behind it.

Fixes #80